requirements.txt
grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.17.3
pymongo==3.12.0
six==1.16.0
